add caching for 5 minutes for each unique request
Some requests are too onerous with several thousand records
being retrieved from the database. Using the combination of
the request parameters as key, we store the results in the
cache, and serve any subsequent queries with the same
parameters with the cached results. The cache is set to
expire after 5 minutes.